all tests passing
waiting on script view PR to be merged then will adjust those and will be done soon!
